server/requirements.txt
-i https://pypi.org/simple
asgiref==3.11.0; python_version >= '3.9'
attrs==25.4.0; python_version >= '3.9'
autobahn==25.10.2; python_version >= '3.11'
automat==25.4.16; python_version >= '3.9'
cffi==2.0.0; python_version >= '3.9'
channels==4.3.2; python_version >= '3.9'
channels-redis==4.3.0; python_version >= '3.9'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
constantly==23.10.4; python_version >= '3.8'
cryptography==46.0.3; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
daphne==4.2.1; python_version >= '3.9'
dj-database-url==3.0.1
django==5.2.8; python_version >= '3.10'
django-cors-headers==4.9.0; python_version >= '3.9'
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt[crypto]==5.5.1; python_version >= '3.9'
hyperlink==21.0.0
idna==3.11; python_version >= '3.8'
incremental==24.7.2; python_version >= '3.8'
iniconfig==2.3.0; python_version >= '3.10'
msgpack==1.1.2; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
pluggy==1.6.0; python_version >= '3.9'
psycopg2-binary==2.9.11; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.23; python_version >= '3.8'
pygments==2.19.2; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
pyopenssl==25.3.0; python_version >= '3.7'
pytest==9.0.1; python_version >= '3.10'
pytest-django==4.11.1; python_version >= '3.8'
python-dotenv==1.2.1; python_version >= '3.9'
redis==7.1.0; python_version >= '3.10'
service-identity==24.2.0; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
twisted[tls]==25.5.0; python_full_version >= '3.8.0'
txaio==25.9.2; python_version >= '3.10'
typing-extensions==4.15.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
zope-interface==8.1.1; python_version >= '3.10'
